requirements.txt
scipy==1.6.2
numpy==1.19.3
cython==0.29.22
h5py==3.2.1 
matplotlib==3.4.1
psd==1.5.2
pyyaml==5.4.1
astropy==4.2
xarray==0.15.0
tables
lisaconstants # see https://gitlab.in2p3.fr/lisa-simulation/constants
h5netcdf
